Forward `Diagonal` `rmul!`/`lmul!` for adj/trans to parent (#1229)
Currently. `rmul!(A'::Adjoint, D::Diagonal)` loops over
`CartesianIndices(A')`, which is not cache-friendly. In this PR, we use
the fact that `A' = A' * D` may be rewritten as `A = D' * A`, and that
`D' isa Diagonal` by design. The operation therefore becomes
`lmul!(D'::Diagonal, A)`, which is cache-friendly.
On master,
```julia
julia> using LinearAlgebra, Chairmarks

julia> D = Diagonal(rand(4000));

julia> A = rand(size(D)...);

julia> @b (A', D) rmul!(_[1], _[2])
103.276 ms
```
whereas, with this PR,
```julia
julia> @b (A', D) rmul!(_[1], _[2])
9.994 ms
```